+++ src/Nethereum.Unity/FeeSuggestions/SuggestTipUnityRequestStrategy.cs
@@ -40,7 +40,7 @@ namespace Nethereum.Unity.FeeSuggestions
                 if (blockCount > 0)
                 {
                     // feeHistory API call with reward percentile specified is expensive and therefore is only requested for a few non-full recent blocks.
-                    yield return _ethFeeHistory.SendRequest(blockCount.ToHexBigInteger(), new BlockParameter(new HexBigInteger(firstBlock + ptr)), new double[] { 0 });
+                    yield return _ethFeeHistory.SendRequest(blockCount.ToHexBigInteger(), new BlockParameter(new HexBigInteger(firstBlock + ptr)), new decimal[] { 0 });
 
                     if (_ethFeeHistory.Exception == null)
                     {
